Validate data argument

Make sure that, if specified, data is a data frame. If data is not missing then dimension reduction is also performed on variables (i.e., a one column matrix inside a data frame is converted to a vector).

drop_redundant_dims(data)

Arguments

data
User's data argument

if_missing
Object to return if data is missing/null

Value

If no error is thrown, data itself is returned if not missing/null, otherwise if_missing is returned.