In-sample or out-of-sample predictive errors

This is a convenience function for computing \(y - y^{rep}\) (in-sample, for observed \(y\)) or \(y - \tilde{y}\) (out-of-sample, for new or held-out \(y\)). The method for stapreg objects calls posterior_predict internally, whereas the method for objects with class "ppd" accepts the matrix returned by posterior_predict as input and can be used to avoid multiple calls to posterior_predict.

# S3 method for stapreg
predictive_error(object, newsubjdata = NULL,
  newdistdata = NULL, draws = NULL, re.form = NULL, seed = NULL,
  offset = NULL, ...)

# S3 method for ppd
predictive_error(object, y, ...)

Arguments

object
Either a fitted model object returned by one of the rstap modeling functions (a stapreg object) or, for the "ppd" method, a matrix of draws from the posterior predictive distribution returned by posterior_predict.

...
Currently ignored.

y
For the "ppd" method only, a vector of \(y\) values the same length as the number of columns in the matrix used as object. The method for stapreg objects takes y directly from the fitted model object.

newdata, draws, seed, offset, re.form
Optional arguments passed to posterior_predict. For binomial models, please see the Note section below if newdata will be specified.

Value

A draws by nrow(newdata) matrix. If newdata is not specified then it will be draws by nobs(object).

Note

The Note section in posterior_predict about newdata for binomial models also applies for predictive_error, with one important difference. For posterior_predict if the left-hand side of the model formula is cbind(successes, failures) then the particular values of successes and failures in newdata don't matter, only that they add to the desired number of trials. This is not the case for predictive_error. For predictive_error the particular value of successes matters because it is used as \(y\) when computing the error.

See also

Examples

if (!exists("example_model")) example(example_model)

#> Warning: no help found for ‘example_model’

err1 <- predictive_error(example_model, draws = 50)

#> Error in predictive_error(example_model, draws = 50): object 'example_model' not found

hist(err1)

#> Error in hist(err1): object 'err1' not found

# Using newdata with a binomial model
formula(example_model)

#> Error in formula(example_model): object 'example_model' not found

nd <- data.frame(
 size = c(10, 20),
 incidence = c(5, 10),
 period = factor(c(1,2)),
 herd = c(1, 15)
)
err2 <- predictive_error(example_model, newdata = nd, draws = 10, seed = 1234)

#> Error in predictive_error(example_model, newdata = nd, draws = 10, seed = 1234): object 'example_model' not found

# stapreg vs ppd methods
fit <- stan_glm(mpg ~ wt, data = mtcars, iter = 300)

#> Error in stan_glm(mpg ~ wt, data = mtcars, iter = 300): could not find function "stan_glm"

preds <- posterior_predict(fit, seed = 123)

#> Error in posterior_predict(fit, seed = 123): object 'fit' not found

all.equal(
  predictive_error(fit, seed = 123),
  predictive_error(preds, y = fit$y)
)

#> Error in predictive_error(fit, seed = 123): object 'fit' not found