Print method for stapreg objects

The print method for stapreg objects displays a compact summary of the fitted model. See the Details section below for descriptions of the different components of the printed output. For additional summary statistics and diagnostics use the summary method.

# S3 method for stapreg
print(x, digits = 1, include_X = FALSE, ...)

Arguments

x
A fitted model object returned by one of the rstap modeling functions. See stapreg-objects.

digits
Number of digits to use for formatting numbers.

include_X
logical for whether or not to include estimated latent exposure covariate

...
Ignored.

Value

Returns x, invisibly.

Details

Point estimates
Point estimates are medians computed from simulations. For models fit using MCMC ("sampling") the posterior sample is used. The point estimates reported are the same as the values returned by coef.
Uncertainty estimates (MAD_SD)
The standard deviations reported (labeled MAD_SD in the print output) are computed from the same set of draws described above and are proportional to the median absolute deviation (mad) from the median. Compared to the raw posterior standard deviation, the MAD_SD will be more robust for long-tailed distributions. These are the same as the values returned by se.
Additional output

The median and MAD_SD are also reported for mean_PPD, the sample average posterior predictive distribution of the outcome. This is useful as a quick diagnostic. A useful heuristic is to check if mean_PPD is plausible when compared to mean(y). If it is plausible then this does not mean that the model is good in general (only that it can reproduce the sample mean), however if mean_PPD is implausible then it is a sign that something is wrong (severe model misspecification, problems with the data, computational issues, etc.).

For GLMs with group-specific terms (see stap_glmer) the printed output also shows point estimates of the standard deviations of the group effects (and correlations if there are both intercept and slopes that vary by group).

See also