Summarize the priors used for an rstap model

The prior_summary method provides a summary of the prior distributions used for the parameters in a given model. In some cases the user-specified prior does not correspond exactly to the prior used internally by rstap (see the sections below). Especially in these cases, but also in general, it can be much more useful to visualize the priors. Visualizing the priors can be done using the posterior_vs_prior function, or alternatively by fitting the model with the prior_PD argument set to TRUE (to draw from the prior predictive distribution instead of conditioning on the outcome) and then plotting the parameters.

# S3 method for stapreg
prior_summary(object, digits = 2, ...)

Arguments

object
A fitted model object returned by one of the rstap modeling functions. See stapreg-objects.

digits
Number of digits to use for rounding.

...
Currently ignored by the method for stapreg objects.

Value

A list of class "prior_summary.stapreg", which has its own print method.

Intercept (after predictors centered)

For rstap modeling functions that accept a prior_intercept argument, the specified prior for the intercept term applies to the intercept after rstap internally centers the predictors so they each have mean zero. The estimate of the intercept returned to the user correspond to the intercept with the predictors as specified by the user (unmodified by rstap), but when specifying the prior the intercept can be interpreted as the expected outcome when the predictors are set to their means.

Adjusted scales

For some models you may see "adjusted scale" in the printed output and adjusted scales included in the object returned by prior_summary. These adjusted scale values are the prior scales actually used by rstap and are computed by adjusting the prior scales specified by the user to account for the scales of the predictors (as described in the documentation for the autoscale argument). To disable internal prior scale adjustments set the autoscale argument to FALSE when setting a prior using one of the distributions that accepts an autoscale argument. For example, normal(0, 5, autoscale=FALSE) instead of just normal(0, 5).

See also