Bayesian spatial-temporal aggregated predictor(STAP) models via Stan

Bayesian inference for linear modeling with regularizing priors on the model parameters that are driven by prior beliefs about \(R^2\), the proportion of variance in the outcome attributable to the predictors. See priors for an explanation of this critical point. stan_glm with family="gaussian" also estimates a linear model with normally-distributed errors and allows for various other priors on

stap_lm(formula, subject_data, bef_data, id_key = NULL, max_distance, subset,
  weights, na.action, model = TRUE, x = FALSE, y = FALSE, z = FALSE,
  singular.ok = TRUE, contrasts = NULL, offset, ...,
  prior = R2(stop("'location' must be specified'")), prior_intercept = NULL,
  adapt_delta = NULL)

Arguments

formula, subset
Same as lm,

na.action, singular.ok, contrasts
Same as lm, but rarely specified.

model, offset, weights
Same as lm, but rarely specified.

x, y
In stan_lm, logical scalars indicating whether to return the design matrix and response vector. In stan_lm.fit or stan_lm.wfit, a design matrix and response vector.

...
Further arguments passed to the function in the rstap to specify iter, chains, cores, refresh, etc.

prior
Must be a call to R2 with its location argument specified or NULL, which would indicate a standard uniform prior for the \(R^2\).

prior_intercept
Either NULL (the default) or a call to normal. If a normal prior is specified without a scale, then the standard deviation is taken to be the marginal standard deviation of the outcome divided by the square root of the sample size, which is legitimate because the marginal standard deviation of the outcome is a primitive parameter being estimated.

Note: The prior distribution for the intercept is set so it applies to the value when all predictors are centered. If you prefer to specify a prior on the intercept without the predictors being auto-centered, then you have to omit the intercept from the formula and include a column of ones as a predictor, in which case some element of prior specifies the prior on it, rather than prior_intercept. Regardless of how prior_intercept is specified, the reported estimates of the intercept always correspond to a parameterization without centered predictors (i.e., same as in glm).

adapt_delta
See the adapt_delta help page for details.

w
Same as in lm.wfit but rarely specified.

Value

A stapreg object is returned for stan_lm.

A stapfit object (or a slightly modified stapfit object) is returned if stan_lm.fit or stan_lm.wfit is called directly.

Details

The stan_lm function is similar in syntax to the lm function but rather than choosing the parameters to minimize the sum of squared residuals, samples from the posterior distribution are drawn using MCMC. The stan_lm function has a formula-based interface and would usually be called by users but the stap_lm.fit and stap_lm.wfit functions might be called by other functions that parse the data themselves and are analgoous to lm.fit and lm.wfit respectively.

In addition to estimating sigma --- the standard deviation of the normally-distributed errors --- this model estimates a positive parameter called log-fit_ratio. If it is positive, the marginal posterior variance of the outcome will exceed the sample variance of the outcome by a multiplicative factor equal to the square of fit_ratio. Conversely if log-fit_ratio is negative, then the model underfits. Given the regularizing nature of the priors, a slight underfit is good.

Finally, the posterior predictive distribution is generated with the predictors fixed at their sample means. This quantity is useful for checking convergence because it is reasonably normally distributed and a function of all the parameters in the model.

References

Lewandowski, D., Kurowicka D., and Joe, H. (2009). Generating random correlation matrices based on vines and extended onion method. Journal of Multivariate Analysis. 100(9), 1989--2001.

See also