Methods for stapreg objects

The methods documented on this page are actually some of the least important methods defined for stapreg objects. The most important methods are documented separately, each with its own page. Links to those pages are provided in the See Also section, below.

# S3 method for stapreg
coef(object, ...)

# S3 method for stapreg
confint(object, ...)

# S3 method for stapreg
fitted(object, ...)

# S3 method for stapreg
nobs(object, ...)

# S3 method for stapreg
nstap(object)

# S3 method for stapreg
ntap(object)

# S3 method for stapreg
nsap(object)

# S3 method for stapreg
nfix(object, ...)

# S3 method for stapreg
residuals(object, ...)

# S3 method for stapreg
se(object, ...)

# S3 method for stapreg
vcov(object, correlation = FALSE, ...)

# S3 method for stapreg
fixef(object, ...)

# S3 method for stapreg
ngrps(object, ...)

# S3 method for stapreg
ranef(object, ...)

# S3 method for stapreg
sigma(object, ...)

# S3 method for stapreg
VarCorr(x, sigma = 1, ...)

Arguments

object, x
A fitted model object returned by one of the rstap modeling functions. See stapreg-objects.

...
Ignored

correlation
For vcov, if FALSE (the default) the covariance matrix is returned. If TRUE, the correlation matrix is returned instead.

sigma
Ignored (included for compatibility with VarCorr).

Details

The methods documented on this page are similar to the methods defined for objects of class 'lm', 'glm', 'glmer', etc. However there are a few key differences:

residuals

Residuals are always of type "response" (not "deviance" residuals or any other type).

coef

Medians are used for point estimates. See the Point estimates section in print.stapreg for more details.

se

The se function returns standard errors based on mad. See the Uncertainty estimates section in print.stapreg for more details.

confint

confint will throw an error because the posterior_interval function should be used to compute Bayesian uncertainty intervals.

See also