Extract the posterior sample

For models fit using MCMC the posterior sample ---the post-warmup draws from the posterior distribution--- can be extracted from a fitted model object as a matrix, data frame, or array. The as.matrix and as.data.frame methods merge all chains together, whereas the as.array method keeps the chains separate.

# S3 method for stapreg
as.matrix(x, ..., pars = NULL, regex_pars = NULL)

# S3 method for stapreg
as.array(x, ..., pars = NULL, regex_pars = NULL)

# S3 method for stapreg
as.data.frame(x, ..., pars = NULL, regex_pars = NULL)

Arguments

x
A fitted model object returned by one of the rstap modeling functions. See stapreg-objects.

...
Ignored.

pars
An optional character vector of parameter names.

regex_pars
An optional character vector of regular expressions to use for parameter selection. regex_pars can be used in place of pars or in addition to pars. Currently, all functions that accept a regex_pars argument ignore it for models fit using optimization.

Value

A matrix, data.frame, or array, the dimensions of which depend on pars and regex_pars, as well as the model and estimation algorithm (see the Description section above).

See also