Datasets for rstap examples

Small datasets for use in rstap examples and vignettes.

Format

homog_longitudinal_bef_data

Simulated data for the longitudinal simulation

subj_ID: The subject unique identifier

measure_ID: The measurement unique identifier

bef_ID The Built Environment Unique identifier

measure_date The date at which the subject was measured

date_open: The date at which the business opened

date_close: The date at which the business may have closed; NA if the business is still open

date: The date at which the subject first moved to the location associated with the distance and time with the built environment feature

class: The kind of built environment feature. Only one is in the simulated dataset - "Coffee Shop"

dist: The distance between the subject and BEF at the date to be associated with the measure ID

time: The time for which the subject was "exposed" to the BEF at corresponding distance

Source: https://biostatistics4socialimpact.github.io/rstap/articles/longitudinal-I.htmlLongitudinal Vignette
homog_longitudinal_subject_data

subj_ID: The subject unique identifier

Income: Simulated continuous covariate

measure_date: The simulated date the subject was measured

ran_int: Random intercept generated for the longitudinal I simulation

y: Continuous outcome simulated for longitudinal I simulation - meant to be akin to BMI

y_bern: Bernoulli outcome simulated

sex: Discrete 1-0 covariate simulated to be akin to sex

Coffee_Shop: The "true" Coffee Shop Exposure covariate

centered_income: scaled and centered version of Income covariate

centered_age: scaled and centered version of Age covariate

Source: \hrefhttps://biostatistics4socialimpact.github.io/rstap/articles/longitudinal-I.htmlLongitudinal Vignette
homog_subject_data

subj_idThe subject unique identifier

yContinuous simulated outcome, meant to be BMI

sexdiscrete factor coded "M" for male, "F" for females

Source: https://biostatistics4socialimpact.github.io/rstap/articles/Introduction.htmlIntroduction Vignette
homog_distance_data

subj_id: The subject unique identifier

BEF Built Environment Feature class identifier - only one included in this dataset "Fast_Food"

Distance: The euclidean distance between the row's subject and Fast Food restaurant locations'

Source: https://biostatistics4socialimpact.github.io/rstap/articles/Introduction.htmlIntroduction Vignette